requirements.txt
pystray[win32]
Pillow
requests
apscheduler
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
httplib2
customtkinter
